requirements.txt
websocket-client>=1.8,<2
matplotlib>=3.8,<4
fastapi>=0.111,<1
uvicorn>=0.30,<1
httpx>=0.27,<1
pydantic>=2.0,<3
